+++ src/factory.c
@@ -6,7 +6,7 @@
 
 #include "factory.h"
 
-void factoryInit(pFactory_t pFactory, size_t numThread, pthreadHandler_t threadHandler)
+void factoryInit(pFactory_t pFactory, int numThread, pthreadHandler_t threadHandler)
 {
     taskQueInit(&pFactory->_que);
     taskQueInit(&pFactory->_queFile);
@@ -19,7 +19,7 @@ void factoryInit(pFactory_t pFactory, size_t numThread, pthreadHandler_t threadH
 void factoryStart(pFactory_t pFactory)
 {
     if (0 == pFactory->_flagStart) {
-        for (size_t idx = 0; idx != pFactory->_numThread; ++idx)
+        for (int idx = 0; idx != pFactory->_numThread; ++idx)
         {
             pthread_create(pFactory->_pThreadId + idx, NULL, pFactory->_pThreadHandler, pFactory);
             printf("pthid = %ld\n", pFactory->_pThreadId[idx]);
