+++ src/taskqueue.h
@@ -15,9 +15,9 @@ typedef struct Node { // 存放网络描述符链式结点
     unsigned short _port;
     char _user[64]; // 用户
     char _path[256]; // 目录
-    size_t _idxLen; // 标识目录尾部的游标/目录长度
-    size_t _inum; // 目录层级
-    size_t _lenDir[16]; // 每级目录的长度（包含/）
+    int _idxLen; // 标识目录尾部的游标/目录长度
+    int _inum; // 目录层级
+    int _lenDir[16]; // 每级目录的长度（包含/）
     char _fileName[64];
     char _flagCmd;
     char _flagSigninStatus;
@@ -28,8 +28,8 @@ typedef struct Node { // 存放网络描述符链式结点
 
 typedef struct { // 存放网络描述符链式队列
     pNode_t _pHead, _pTail;
-    size_t _capacity; // 队列的最大容量
-    size_t _size; // 当前队列大小
+    int _capacity; // 队列的最大容量
+    int _size; // 当前队列大小
     pthread_mutex_t _mutex; // 队列锁，保证入队和出队安全
 } Que_t, * pQue_t;
